refactor: created a extension method on the testrun, this method will map the time, if the test succeeded to the right feature.
Refs: #4
